apps/llm/requirements.txt
fastapi
uvicorn
sentence-transformers
pandas
numpy
openai
python-dotenv
validators
GitPython
python-multipart